extract ChapterContentsLoader
