+++ general/book-assert/BookLoader/index.ts
@@ -1,9 +1,4 @@
-import {
-  ChapterHeading,
-  PageHeading,
-  ReadmeHeading,
-  Summary,
-} from "../summary";
+import { ChapterHeading, ReadmeHeading, Summary } from "../summary";
 import { DirectoryPath } from "../../file-paths/DirectoryPath";
 import { FilePath } from "../../file-paths/FilePath";
 import { promisify } from "util";
@@ -12,7 +7,7 @@ import * as md from "markdown-it";
 import * as cheerio from "cheerio";
 import { BookChapters, ChapterContent } from "../BookChapter";
 import { ChapterLoader } from "../BookIndexer";
-import { basename } from "path";
+import { ChapterContentsLoader } from "./ChapterContentsLoader";
 
 export class BookLoader {
   constructor(
@@ -30,34 +25,14 @@ export class BookLoader {
   }
 
   async loadBookChapters(): Promise<BookChapters> {
-    const promises = this.chapterPaths.map((path, i) =>
-      this.loadChapterContent(path, i),
-    );
-    const contents = await promises.reduce(async (x1, x2) => {
-      return (await x1).concat(await x2);
-    });
-    return { contents };
-  }
+    const promises = this.chapterPaths
+      .map(ChapterContentsLoader.from(this.loadChapter))
+      .map(_ => _.loadContents());
 
-  private async loadChapterContent(
-    chapterPath: DirectoryPath,
-    chapterIndex: number,
-  ): Promise<ChapterContent[]> {
-    const f = async (name: string) => {
-      const filePath = await chapterPath.withParent.confirmFile(name);
-      const markdownString = await promisify(readFile)(
-        filePath.toAbsolute,
-        "utf-8",
-      );
-      return {
-        filePath,
-        markdownString: warning + "\n" + markdownString,
-      };
-    };
-    const nextPages = await this.loadNextPages(chapterPath, chapterIndex + 1);
-    return Promise.all(
-      [f("index.md")].concat(nextPages.map(page => f(basename(page.path)))),
+    const contents = await Promise.all(promises).then(_ =>
+      _.reduce((a, b) => a.concat(b)),
     );
+    return { contents };
   }
 
   private async loadReadmeHeading(): Promise<ReadmeHeading> {
@@ -66,52 +41,12 @@ export class BookLoader {
 
   private async loadChapterHeadings(): Promise<ChapterHeading[]> {
     console.log("loadChapters", this.root, this.chapterPaths);
-    return Promise.all(this.chapterPaths.map(this.loadChapterHeading));
-  }
+    const promises = this.chapterPaths
+      .map(ChapterContentsLoader.from(this.loadChapter))
+      .map(_ => _.loadChapterHeading());
 
-  private async loadNextPages(
-    chapterPath: DirectoryPath,
-    chapterNumber: number,
-  ): Promise<PageHeading[]> {
-    const chapter = await this.loadChapter(chapterNumber);
-    if (!chapter) {
-      return Promise.resolve([]);
-    }
-    const promises = chapter.nextPages.map(async name => {
-      const filePath = await chapterPath.withParent.confirmFile(name);
-      const fromHtml = (html: string) => {
-        const $ = cheerio.load(html);
-        return {
-          title: $("h1").text(),
-          path: filePath.toRelative,
-        };
-      };
-      return promisify(readFile)(filePath.toAbsolute, "utf-8")
-        .then(fromMarkdown)
-        .then(fromHtml);
-    });
     return Promise.all(promises);
   }
-
-  private loadChapterHeading = async (
-    chapterPath: DirectoryPath,
-    chapterIndex: number,
-  ) => {
-    const index = await chapterPath.withParent.confirmFile("index.md");
-    const nextPages = await this.loadNextPages(chapterPath, chapterIndex + 1);
-    const fromHtml = (html: string) => {
-      const $ = cheerio.load(html);
-      const h1 = $("h1").text();
-      return {
-        title: `${chapterIndex + 1}. ${h1}`,
-        path: index.toRelative,
-        nextPages,
-      };
-    };
-    return promisify(readFile)(index.toAbsolute, "utf-8")
-      .then(fromMarkdown)
-      .then(fromHtml);
-  };
 }
 
 const fromMarkdown = (markdown: string) => md().render(markdown);
@@ -128,19 +63,3 @@ function loadReadme(path: FilePath): Promise<ReadmeHeading> {
     .then(fromMarkdown)
     .then(fromHtml);
 }
-
-function toAnchor(text: string): string {
-  return text
-    .replace(/[.:]/g, "")
-    .replace(/[ ]/g, "-")
-    .toLowerCase();
-}
-
-const warning = `
-{% if false %}
-
-* This file is automatically generated by book-assert library.
-* Do not modify this file -- YOUR CHANGES WILL BE ERASED!
-
-{% endif %}
-`;
